requirements.txt
PyGithub==1.44.1
requests==2.22.0
